# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "myExtBot API",
    "version": "0.1.0",
    "description": "REST API for myExtBot — an extensible multi-agent orchestration platform.\n\nAll endpoints under `/api/*` require an `x-api-key` header (set `MYEXTBOT_API_KEY` env var, default: `dev-key`).\n\nInteractive docs: `/api-docs` · Raw spec: `/api-docs/json`"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "SuccessOk": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "PluginSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "installed": {
            "type": "boolean"
          },
          "installedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/plugins": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "List all plugins (PluginManager — legacy)",
        "responses": {
          "200": {
            "description": "Array of plugin summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginSummary",
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/api/plugins/installed": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "List installed plugins (PluginManager — legacy)",
        "responses": {
          "200": {
            "description": "Array of installed plugin summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginSummary",
                    "type": "object"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/api/plugins/{id}": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Get details for a single plugin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginSummary"
                }
              }
            }
          },
          "404": {
            "description": "Plugin not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/api/plugins/{id}/install": {
      "post": {
        "tags": [
          "Plugins"
        ],
        "summary": "Install a plugin by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin installed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "plugin": {
                      "$ref": "#/components/schemas/PluginSummary"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plugin not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/api/plugins/{id}/uninstall": {
      "delete": {
        "tags": [
          "Plugins"
        ],
        "summary": "Uninstall a plugin by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin uninstalled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plugin not installed"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/api/agents/route": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Route a query to the most relevant agents",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Natural language query to route"
          },
          {
            "in": "query",
            "name": "topN",
            "schema": {
              "type": "integer",
              "default": 3
            },
            "description": "Number of top suggestions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of agent routing suggestions (sorted by score)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "agentId": {
                        "type": "string"
                      },
                      "score": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plugins/{pluginId}": {
      "delete": {
        "tags": [
          "Plugins"
        ],
        "summary": "Uninstall a plugin (PluginManager — legacy)",
        "parameters": [
          {
            "in": "path",
            "name": "pluginId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plugin uninstalled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessOk"
                }
              }
            }
          },
          "400": {
            "description": "Uninstall error"
          }
        }
      }
    },
...
  }
}